Calculate- can add two double number

+++ src/main/com/thoughtworks/lesson/one/Calculator.java
@@ -4,7 +4,7 @@ package com.thoughtworks.lesson.one;
  * Created by Thoughtworks on 5/24/15.
  */
 public class Calculator {
-    public int add(int first, int second) {
+    public double add(double first, double second) {
         return first+second;
     }
 }
